Clean up structure.
